server/pyproject.toml
[tool.poetry]
name = "mezza"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "mezza" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg = "^3.1.16"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
sentry-sdk = {extras = ["django"], version = "^2.3.1"}
isort = "^5.13.2"
django-bridge = "^0.3"
whitenoise = "^6.6.0"
pillow = "^10.3.0"
filetype = "^1.2.0"
django-storages = {extras = ["s3"], version = "^1.14.4"}
ffmpeg-python = "^0.2.0"
pypdf2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.5.5"

[tool.isort]
profile = "black"
known_first_party = ["mezza"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
